chore: Remove unused code and optimize imports in SettingScreen.kt, CustomCameraRepoImpl.kt, CameraViewModel.kt, MainActivity.kt, and MainScreen.kt
