+++ app/src/main/java/com/kamatiakash/customcameraappwithjetpackcompose/MainActivity.kt
@@ -3,15 +3,8 @@ package com.kamatiakash.customcameraappwithjetpackcompose
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Surface
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
 import com.kamatiakash.customcameraappwithjetpackcompose.presentation.MainScreen
-import com.kamatiakash.customcameraappwithjetpackcompose.ui.theme.CustomCameraAppWithJetpackComposeTheme
+import com.kamatiakash.customcameraappwithjetpackcompose.ui.theme.SimpleNavComposeAppTheme
 import dagger.hilt.android.AndroidEntryPoint
 
 
@@ -20,13 +13,8 @@ class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-            CustomCameraAppWithJetpackComposeTheme {
-                Surface(
-                    modifier = Modifier.fillMaxSize(),
-                    color = MaterialTheme.colors.background
-                ) {
-                    MainScreen()
-                }
+            SimpleNavComposeAppTheme {
+                MainScreen()
             }
         }
     }
