+++ app/src/main/java/com/kamatiakash/customcameraappwithjetpackcompose/presentation/MainScreen.kt
@@ -19,91 +19,16 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.viewinterop.AndroidView
 import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.navigation.compose.rememberNavController
 import com.google.accompanist.permissions.ExperimentalPermissionsApi
 import com.google.accompanist.permissions.rememberMultiplePermissionsState
 import com.kamatiakash.customcameraappwithjetpackcompose.R
+import com.kamatiakash.customcameraappwithjetpackcompose.presentation.navigation.NavGraph
 
 
 @OptIn(ExperimentalPermissionsApi::class)
 @Composable
-fun MainScreen(
-    viewModel: MainViewModel = hiltViewModel()
-) {
-
-    val permissions = if (Build.VERSION.SDK_INT <= 28){
-        listOf(
-            Manifest.permission.CAMERA,
-            Manifest.permission.WRITE_EXTERNAL_STORAGE
-        )
-    }else listOf(Manifest.permission.CAMERA)
-
-    val permissionState = rememberMultiplePermissionsState(
-        permissions = permissions)
-
-    if (!permissionState.allPermissionsGranted){
-        SideEffect {
-            permissionState.launchMultiplePermissionRequest()
-        }
-    }
-
-
-    val context = LocalContext.current
-    val lifecycleOwner = LocalLifecycleOwner.current
-    val configuration = LocalConfiguration.current
-    val screeHeight = configuration.screenHeightDp.dp
-    val screenWidth = configuration.screenWidthDp.dp
-    var previewView:PreviewView
-
-
-    Column(
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center
-    ) {
-        // we will show camera preview once permission is granted
-        if (permissionState.allPermissionsGranted){
-            Box(modifier = Modifier
-                .height(screeHeight * 0.85f)
-                .width(screenWidth)) {
-                AndroidView(
-                    factory = {
-                        previewView = PreviewView(it)
-                        viewModel.showCameraPreview(previewView, lifecycleOwner)
-                        previewView
-                    },
-                    modifier = Modifier
-                        .height(screeHeight * 0.85f)
-                        .width(screenWidth)
-                )
-            }
-        }
-
-        Box(
-            modifier = Modifier
-                .height(screeHeight*0.15f),
-            contentAlignment = Alignment.Center
-        ){
-            IconButton(onClick = {
-                if (permissionState.allPermissionsGranted){
-                    viewModel.captureAndSave(context)
-                }
-                else{
-                    Toast.makeText(
-                        context,
-                        "Please accept permission in app settings",
-                        Toast.LENGTH_LONG
-                    ).show()
-                }
-            }) {
-
-                Icon(painter =
-                painterResource(id =
-                R.drawable.ic_baseline_camera_alt_24),
-                    contentDescription = "",
-                    modifier = Modifier.size(45.dp),
-                    tint = Color.Magenta
-                )
-
-            }
-        }
-    }
+fun MainScreen() {
+    val navController = rememberNavController()
+    NavGraph(navController)
 }
