+++ app/src/main/java/com/kamatiakash/customcameraappwithjetpackcompose/presentation/screen/CameraScreen.kt
@@ -1,5 +1,6 @@
 package com.kamatiakash.customcameraappwithjetpackcompose.presentation.screen
 
+import android.Manifest
 import android.widget.Toast
 import androidx.camera.view.PreviewView
 import androidx.compose.foundation.layout.Arrangement
@@ -11,6 +12,7 @@ import androidx.compose.foundation.layout.width
 import androidx.compose.material.Icon
 import androidx.compose.material.IconButton
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.SideEffect
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -21,10 +23,13 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.viewinterop.AndroidView
 import androidx.hilt.navigation.compose.hiltViewModel
+import com.google.accompanist.permissions.ExperimentalPermissionsApi
+import com.google.accompanist.permissions.rememberMultiplePermissionsState
 import com.kamatiakash.customcameraappwithjetpackcompose.R
 import com.kamatiakash.customcameraappwithjetpackcompose.presentation.MainViewModel
 import com.kamatiakash.customcameraappwithjetpackcompose.presentation.viewmodel.CameraViewModel
 
+@OptIn(ExperimentalPermissionsApi::class)
 @Composable
 fun CameraScreen(
     navigationFrame: () -> Unit,
@@ -43,38 +48,76 @@ fun CameraScreen(
         horizontalAlignment = Alignment.CenterHorizontally,
         verticalArrangement = Arrangement.Center
     ) {
-        // we will show camera preview once permission is granted
-            Box(modifier = Modifier
-                .height(screeHeight * 0.85f)
-                .width(screenWidth)) {
-                AndroidView(
-                    factory = {
-                        previewView = PreviewView(it)
-                        viewModel.showCameraPreview(previewView, lifecycleOwner)
-                        previewView
-                    },
-                    modifier = Modifier
-                        .height(screeHeight * 0.85f)
-                        .width(screenWidth)
-                )
+
+        val permissions = listOf(Manifest.permission.CAMERA)
+
+        val permissionState = rememberMultiplePermissionsState(
+            permissions = permissions)
+
+        if (!permissionState.allPermissionsGranted){
+            SideEffect {
+                permissionState.launchMultiplePermissionRequest()
+            }
+        }
+
+
+        val context = LocalContext.current
+        val lifecycleOwner = LocalLifecycleOwner.current
+        val configuration = LocalConfiguration.current
+        val screeHeight = configuration.screenHeightDp.dp
+        val screenWidth = configuration.screenWidthDp.dp
+        var previewView:PreviewView
+
+
+        Column(
+            horizontalAlignment = Alignment.CenterHorizontally,
+            verticalArrangement = Arrangement.Center
+        ) {
+            // we will show camera preview once permission is granted
+            if (permissionState.allPermissionsGranted){
+                Box(modifier = Modifier
+                    .height(screeHeight * 0.85f)
+                    .width(screenWidth)) {
+                    AndroidView(
+                        factory = {
+                            previewView = PreviewView(it)
+                            viewModel.showCameraPreview(previewView, lifecycleOwner)
+                            previewView
+                        },
+                        modifier = Modifier
+                            .height(screeHeight * 0.85f)
+                            .width(screenWidth)
+                    )
+                }
             }
 
-        Box(
-            modifier = Modifier
-                .height(screeHeight*0.15f),
-            contentAlignment = Alignment.Center
-        ){
-            IconButton(onClick = {
-            }) {
+            Box(
+                modifier = Modifier
+                    .height(screeHeight*0.15f),
+                contentAlignment = Alignment.Center
+            ){
+                IconButton(onClick = {
+                    if (permissionState.allPermissionsGranted){
+                        viewModel.captureAndSave(context)
+                    }
+                    else{
+                        Toast.makeText(
+                            context,
+                            "Please accept permission in app settings",
+                            Toast.LENGTH_LONG
+                        ).show()
+                    }
+                }) {
 
-                Icon(painter =
-                painterResource(id =
-                R.drawable.ic_baseline_camera_alt_24),
-                    contentDescription = "",
-                    modifier = Modifier.size(45.dp),
-                    tint = Color.Magenta
-                )
+                    Icon(painter =
+                    painterResource(id =
+                    R.drawable.ic_baseline_camera_alt_24),
+                        contentDescription = "",
+                        modifier = Modifier.size(45.dp),
+                        tint = Color.Magenta
+                    )
 
+                }
             }
         }
     }
