+++ app/src/main/java/com/kamatiakash/customcameraappwithjetpackcompose/presentation/viewmodel/CameraViewModel.kt
@@ -1,12 +1,14 @@
 package com.kamatiakash.customcameraappwithjetpackcompose.presentation.viewmodel
 
 import android.content.Context
+import android.os.CountDownTimer
 import androidx.camera.view.PreviewView
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.kamatiakash.customcameraappwithjetpackcompose.domain.repository.CustomCameraRepo
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
@@ -31,6 +33,21 @@ class CameraViewModel @Inject constructor(
         viewModelScope.launch {
             repo.captureAndSaveImage(context)
         }
+    }
+
+    fun timerCapture(context: Context, time : Long){
+            viewModelScope.launch {
+                delay(1000L)
+                val counter = object : CountDownTimer(time, 5000) {
+                    override fun onTick(millisUntilFinished: Long) {
+                        captureAndSave(context)
+                    }
+
+                    override fun onFinish() {
+                        this.cancel()
+                    }
+                }
+            }
     }
 
 }
